api/requirements.txt
django==3.1.7
djangorestframework==3.12.2
mysqlclient==2.0.3
python-dotenv==0.15.0
django-cors-headers==3.5.0
django-filter==2.4.0
Pillow==8.2.0
djoser==2.1.0
django-storages==1.11.1
boto3==1.17.94